+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Kiss the girls".
 - Tests for "Random walk".
 - Tests for "Shadow casting".
+- Tests for "Fire control".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".